weekly payroll report.
Write a program that displays a weekly payroll report.
A loop in the program should ask the user for the employee number, gross pay, state tax, federal tax, and FICA withholdings.
The loop will terminate when 0 is entered for the employee number. 
After the data is entered, the program should display totals for gross pay, state tax, federal tax, FICA withholdings and net pay.
Validation: Do not accept negative numbers for any of the items entered. 
Do not accept values for the state, federal or FICA that are greater than the gross pay.
If the sum state tax + FICA for any employee is greater than gross pay, 
print an error message and ask the user to re-enter the data for that employee.