The distance a vehicle travels can be calculated
/*
The distance a vehicle travels can be calculated as follows:
distance = speed * time
For example, if a train travels 40 miles per hour for 3 hours, the distance traveled is 120 miles.
Write a program that asks the user for the speed of a vehicle and how many hours it has traveled. 
The program should then use a loop to display the distance the vehicle traveled for each hour of that time. 

*/