Sum of Values
Sum of Values
Write a program that asks the user for a positive integer value.
The program should use a loop to get the sum of all the integers from 1 up to the number entered. 
For example, if the user enters 50, the loop will sum 1+2+3+...+50.
Input validation: Do not accept negative numbers.