Pennies for a Work
Write a program that calculates how much a person  would earn over a period of time if his or her
salary is one penny the first day, two pennies the second day, and continues to double each day.
The program should ask the user for the number of days. Display a table showing how much the salary
was for each day, and then show the total pay at the end of the period. The outputs should be displayed
in a dollar amount, not the number of pennies