Calculating Hotel Occupancy
Write a program that calculates the occupancy rate for a hotel. The program should start by asking user how many
floors the hotel has. A loop should then iterate once for each floor. In each iteration, the loop should ask the user 
for the number of rooms on the floor and how many of them are occupied. After all the iterations, the program should 
display how many rooms the hotel has, how many of them are occupied, how many are unoccupied, and the percentage of 
rooms that are occupied. The percentage may be calculated by dividing the number of rooms occupied by the number of rooms.
Input validation: 

Do not accept a value less than 1 for the number of floors. Do not accept a number less than 10 for
the number of rooms on a floor.

NOTE: It is traditional that most hotels do not have 13th floor.
The loop in this program should skip the entire 13th iteration.