Population Size & Growth (C++ While and For Loops)
Write a program that will predict the size of a population of organisms.
The program should ask the user for the starting number of organisms,
their average daily population increase (as a percentage), 
and the number of days they will multiply.
A loop should display the size of the population for each day.
Input Validation: 
Do not accept a number less than 2 for the starting size of the population. 
Do not accept a negative number for average daily population increase.
Do not accept a number less than 1 for the number of days they will multiply.