Finding Lowest & Highest String Using C++ Loops
A teacher asked all her students to line up single file according to their first name.
For example, Amy will be at the front and Yolanda at the end. Write a program that prompts the user to
enter the numbers of students in the class, then loops to read in that many names. 
Once all the names have been read in it reports which student would be at the front of the line and which one would be at the end. Assume that no two students have the same name.
Input Validation: Do not accept a number less than 1 or greater than 25 for the numbers of students.