Guess The Number Game - Using Loops and If
Write a program that generates a random number and asks the user to guess what the number is.
If the user's guess is higher than the random number, the program should display "Too high, try again". 
If it is lower, it should display "Too low, try again". 
The program should use a loop that repeats until the user correctly guesses the random number.