Display Square Pattern Using Nested For Loops
Write a program that asks the user for a positive integer no greater than 15. 
The program should then display a square on the screen using the character 'X'. 
The number entered by the user will be the length of each side of the square. 
So if user enters 5, we will display

XXXXX
XXXXX
XXXXX
XXXXX
XXXXX