Calculate Profits Using Functions and Switch
Write a program that determines which of a company's four division (NE, SE, NW, SW) had the greatest sales for a quarter. 
It should include the following two functions, which are called by Main:
- double getSales() is passed the name of a division.
It asks the user for a division's quarterly sales figure, validates the input, then returns it. It should be called once for each division.
- void findHighest() is passed the four sales totals. It determines which is the largest and prints the name if the high with its sale figure.
Input Validation: 
Do not accept dollar amounts less than $0.00