Calculate Values Using Functions and Switch
Write a program that determines which of five geographic regions withing a major city 
(north, south, east, west, central) had the fewest reported automobile accidents last year.
It should have the following functions, which are called by main:
- int getNumAccidents() is passed the name of a region.
It asks the user for the number of automobile accidents reported in that region during the last year, 
validates the input, then returns it. It should be called once for each city region.
- void findLowest() is passed the five accident totals. 
It determines which is the smallest and prints the name of the region, along with its accident figure.
Input Validation: Do not accept negative values.