String Comparison (String.Compare, String Methods)
Write a program that asks the user to enter two names and store them in string objects.
It should then report whether they are the same or not, ignoring case, To help accomplish its task, 
it should use two functions in addition to main(). The functions are:
StringUppercaseIt(string s) boolean sameString(string s1,string s2)
The sameString function, which receives the two strings to be compared, 
will need to call uppercaseIt for each of them before testing if they are the same.
The uppercaseIt function should use a loop so that it can call the to upper function 
for every character in the string it receives before returning it back to the sameString function.